minor changes (relocation) to paged-result module.
